fix: MythicMobs Triggers & Effects
