+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/mythicmobs/MythicMobsIntegration.kt
@@ -3,9 +3,9 @@ package com.willfp.libreforge.integrations.mythicmobs
 import com.willfp.eco.core.EcoPlugin
 import com.willfp.libreforge.effects.Effects
 import com.willfp.libreforge.integrations.LoadableIntegration
-import com.willfp.libreforge.integrations.mythicmobs.impl.EffectCastMythicSkill
-import com.willfp.libreforge.integrations.mythicmobs.impl.EffectMythicTelekinesis
-import com.willfp.libreforge.integrations.mythicmobs.impl.TriggerTakeMythicDamage
+import com.willfp.libreforge.integrations.mythicmobs.impl.effect.EffectCastMythicSkill
+import com.willfp.libreforge.integrations.mythicmobs.impl.effect.EffectMythicTelekinesis
+import com.willfp.libreforge.integrations.mythicmobs.impl.trigger.TriggerTakeMythicDamage
 import com.willfp.libreforge.triggers.Triggers
 
 object MythicMobsIntegration : LoadableIntegration {
