+++ core/common/src/main/kotlin/com/willfp/libreforge/triggers/impl/TriggerTakeDamage.kt
@@ -1,12 +1,11 @@
 package com.willfp.libreforge.triggers.impl
 
+import com.willfp.libreforge.integrations.mythicmobs.utils.isMythicMob
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
 import com.willfp.libreforge.triggers.tryAsLivingEntity
-import io.lumine.mythic.bukkit.MythicBukkit
-import org.bukkit.Bukkit
 import org.bukkit.entity.LivingEntity
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
@@ -29,24 +28,10 @@ object TriggerTakeDamage : Trigger("take_damage") {
 
     @EventHandler(ignoreCancelled = true)
     fun handle(event: EntityDamageEvent) {
-
+        if (event.cause in ignoredCauses) return
+        // If Damager (or Projectile Shooter) is MythicMob, then skip. Use 'take_mythic_damage' instead.
+        if (event is EntityDamageByEntityEvent && event.damager.tryAsLivingEntity()?.isMythicMob() == true) return
         val victim = event.entity
-
-        if (event.cause in ignoredCauses) {
-            return
-        }
-
-        if (Bukkit.getPluginManager().isPluginEnabled("MythicMobs")) {
-            if (event is EntityDamageByEntityEvent) {
-                val attacker = event.damager.tryAsLivingEntity()
-                if (attacker != null) {
-                    if (MythicBukkit.inst().mobManager.isMythicMob(attacker)) {
-                        return
-                    }
-                }
-            }
-        }
-
         this.dispatch(
             victim.toDispatcher(),
             TriggerData(
